fix(MovieAPI): category query refactoring.
