+++ BAZProject/MovieAPI/MovieAPI.swift
@@ -21,26 +21,36 @@ class MovieAPI {
     /// Returns the movies list of the category given in Mexico region and in spanish laguague.
     ///
     ///  - Parameter category: The category to be consulted
-    ///  - Returns: [Movie]
+    ///  - Returns: T object.
     ///
     
-    func getMoviesBy(category: MovieAPICategory, completionHandler: @escaping ([Movie]?, Error?) -> Void) {
-        if let url  = URL(string: "https://api.themoviedb.org/3/\(category.endpointUrl)?api_key=\(apiKey)&\(language)&\(region)") {
-            let task = URLSession.shared.dataTask(with: url) { data, response, error in
+    struct URLSessionFetcher {
+        
+        private let urlRequestFactory: URLRequestFactory
+        private let decodableResultAdapter: DecodableResultAdapter
+        private struct DataNotFoundError: Error { }
+        
+        init(urlRequestFactory: URLRequestFactory, decodableResultAdapter: DecodableResultAdapter) {
+            self.urlRequestFactory = urlRequestFactory
+            self.decodableResultAdapter = decodableResultAdapter
+        }
+        
+        func fetchData<T: Decodable>(completionHandler: @escaping (T?, Error?) -> Void) {
+            let urlRequest = urlRequestFactory.makeUrlRequest()
+            
+            let task = URLSession.shared.dataTask(with: urlRequest) { data, response, error in
                 if let error = error {
                     completionHandler(nil, error)
-                }else{
-                    if let data = data,
-                       let json = try? JSONDecoder().decode(MovieAPIResult.self, from: data){
-                        DispatchQueue.main.async {
-                            completionHandler(json.results, nil)
-                        }
+                } else {
+                    guard let data = data
+                    else {
+                        completionHandler(nil, DataNotFoundError()); return
                     }
+                    let movieResult: (T?, Error?) = self.decodableResultAdapter.mapToResult(with: data)
+                    completionHandler(movieResult.0, movieResult.1)
                 }
             }
             task.resume()
-        } else {
-            completionHandler(nil, nil)
         }
     }
